// packages/database/prisma-generated/zod/inputTypeSchemas/InvitationCreateInputSchema.ts
import type { Prisma } from '../../client';

import { z } from 'zod';
import { OrganizationCreateNestedOneWithoutInvitationsInputSchema } from './OrganizationCreateNestedOneWithoutInvitationsInputSchema';
import { TeamCreateNestedOneWithoutInvitationsInputSchema } from './TeamCreateNestedOneWithoutInvitationsInputSchema';
import { UserCreateNestedOneWithoutInvitationsSentInputSchema } from './UserCreateNestedOneWithoutInvitationsSentInputSchema';

export const InvitationCreateInputSchema: z.ZodType<Prisma.InvitationCreateInput> = z.object({
  id: z.string(),
  email: z.string(),
  role: z.string(),
  status: z.string(),
  expiresAt: z.coerce.date(),
  createdAt: z.coerce.date(),
  updatedAt: z.coerce.date().optional().nullable(),
  organization: z.lazy(() => OrganizationCreateNestedOneWithoutInvitationsInputSchema),
  team: z.lazy(() => TeamCreateNestedOneWithoutInvitationsInputSchema).optional(),
  invitedBy: z.lazy(() => UserCreateNestedOneWithoutInvitationsSentInputSchema)
}).strict();

export default InvitationCreateInputSchema;